+++ app/datatables/category_results_datatable.rb
@@ -5,9 +5,9 @@ class CategoryResultsDatatable < IndexDatatable
     columns([:competition_name, :competition_class, :competition_type, :category, :season,
              :ranking, :skater_name, :nation, :points,
              :short_ranking,
-             :short_tss, :short_tes, :short_pcs, :short_deductions, :short_bv,
+             :short_tss, :short_tes, :short_pcs, :short_deductions, :short_base_value,
              :free_ranking,
-             :free_tss, :free_tes, :free_pcs, :free_deductions, :free_bv,
+             :free_tss, :free_tes, :free_pcs, :free_deductions, :free_base_value,
             ])
     
     columns.sources = {