+++ app/updaters/competition_updater.rb
@@ -38,8 +38,16 @@ class CompetitionUpdater
           next unless Category.accept?(category)
 
           update_category_result(competition, category, cat_item[:result_url])
+
           parsed[:segments][category].each do |segment, seg_item|
-            update_score(competition, category, segment, seg_item[:score_url], seg_item[:result_url], segment_starting_time: seg_item[:time])
+            ###
+            ps = competition.performed_segments.create do |ps|
+              ps.category = category
+              ps.segment = segment
+              ps.starting_time = seg_item[:time]
+            end
+            ###
+            update_score(competition, category, segment, seg_item[:score_url], seg_item[:result_url], date: ps.starting_time.to_date)
           end
         end
       end
@@ -68,33 +76,35 @@ class CompetitionUpdater
     end
   end 
   ################
-  def update_score(competition, category, segment, score_url, result_url, segment_starting_time: nil)
+  def update_score(competition, category, segment, score_url, result_url, additionals = {})
     segment_results = @parser.parse_segment_result(result_url)
 
     @parser.parse_score(score_url).each do |sc_parsed|
       competition.scores.create!(category: category, segment: segment) do |score|
         ActiveRecord::Base.transaction {
           ## find skater
-          h = segment_results.select {|h| h[:starting_number] == sc_parsed[:starting_number] }.first ## TODO: for nil
+          h = segment_results.select {|h| h[:starting_number] == sc_parsed[:starting_number] }.first || {} 
 
           skater = ((h[:isu_number].present? ) ?
                      Skater.where(isu_number: h[:isu_number]).first :
                      Skater.where(name: h[:skater_name]).first ) || raise("no such skater")
 
-          ## find relevant category result
+          ## relevant cr
           segment_type = (segment =~ /SHORT/) ? :short : :free
           relevant_cr = competition.category_results.where(category: category, "#{segment_type}_ranking": sc_parsed[:ranking]).first
-          
+          #relevant_cr.update(segment_type => score)
+
           ## set attributes
           score.attributes = {
             category_result: relevant_cr,
             skater: skater,
-            segment_starting_time: segment_starting_time,
-          }
+            #segment_starting_time: segment_starting_time,
+          }.merge(additionals)
           attrs = score.class.column_names.map(&:to_sym) & sc_parsed.keys
           score.attributes = sc_parsed.slice(*attrs)
 
           score.save!  ## need to save here to create children
+
           
           sc_parsed[:elements].map {|e| score.elements.create(e)}
           sc_parsed[:components].map {|e| score.components.create(e)}
@@ -102,30 +112,6 @@ class CompetitionUpdater
           score.update(elements_summary: score.elements.map(&:name).join('/'))
           score.update(components_summary: score.components.map(&:value).join('/'))
           puts score.summary if @verbose
-=begin
-              ## judge details
-              if self.start_date > Time.zone.parse("2016-7-1") # was random order in the past
-                ### elements
-                score.elements.each do |element|
-                  element.judges.split(/\s/).each_with_index do |value, i|
-                    #next if panels[:judges].count <= i+1
-                    element.element_judge_details.create(panel_name: panels[:judges][i+1][:name],
-                                                         panel_nation: panels[:judges][i+1][:nation],
-                                                         number: i, value: value)
-                  end
-                end
-
-                ### component
-                score.components.each do |component|
-                  component.judges.split(/\s/).each_with_index do |value, i|
-                    #next if panels[:judges].count <= i+1
-                    component.component_judge_details.create(panel_name: panels[:judges][i+1][:name],
-                                                             panel_nation: panels[:judges][i+1][:nation],
-                                                             number: i, value: value)
-                  end
-                end
-              end
-=end
         }
       end
     end