+++ db/migrate/003_create_scores.rb
@@ -7,8 +7,8 @@ class CreateScores < ActiveRecord::Migration[5.1]
 
       t.string :category
       t.string :segment
-      #t.date :date, default: Date.new(1970, 1, 1)
-      t.datetime :segment_starting_time, default: Time.new(1970, 1, 1, 0, 0, 0)
+      t.string :segment_type
+      t.date :date, default: Date.new(1970, 1, 1)
       t.string :result_pdf
       
       t.float :tss, default: 0.0